# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Bloom Learning App API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/me:
    put:
      operationId: updateMe
      tags: [users]
      summary: Update current user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserBody"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /users/me/coins:
    post:
      operationId: earnCoins
      tags: [users]
      summary: Earn coins after completing an exercise
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EarnCoinsBody"
      responses:
        "200":
          description: Updated coin balance
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CoinsResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /users/me/claim-gift:
    post:
      operationId: claimMonthlyGift
      tags: [users]
      summary: Claim monthly gift coins
      responses:
        "200":
          description: Gift claimed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClaimGiftResponse"
        "400":
          description: Gift already claimed this month
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /users/me/streak:
    post:
      operationId: checkInStreak
      tags: [users]
      summary: Check in for daily streak
      responses:
        "200":
          description: Streak info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StreakResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /lessons/generate:
    post:
      operationId: generateLesson
      tags: [lessons]
      summary: Generate an AI lesson
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateLessonBody"
      responses:
        "200":
          description: Generated lesson
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lesson"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /lessons/tts:
    post:
      operationId: textToSpeech
      tags: [lessons]
      summary: Convert text to speech for listening exercises
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TTSBody"
      responses:
        "200":
          description: Audio data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TTSResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /lessons/transcribe:
    post:
      operationId: transcribeSpeech
      tags: [lessons]
      summary: Transcribe speech to text for speaking exercises
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TranscribeBody"
      responses:
        "200":
          description: Transcription
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TranscribeResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        displayName:
          type: string
        coins:
          type: integer
        streakCount:
          type: integer
        lastActivityDate:
          type: string
          nullable: true
        lastGiftDate:
          type: string
          nullable: true
        avatarData:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - username
        - displayName
        - coins
        - streakCount
        - lastActivityDate
        - lastGiftDate
        - avatarData
        - createdAt
    UpdateUserBody:
      type: object
      properties:
        displayName:
          type: string
          minLength: 1
          maxLength: 30
        avatarData:
          type: string
      required: []
    EarnCoinsBody:
      type: object
      properties:
        amount:
          type: integer
          minimum: 0
      required:
        - amount
    CoinsResponse:
      type: object
      properties:
        coins:
          type: integer
      required:
        - coins
    ClaimGiftResponse:
      type: object
      properties:
        coins:
          type: integer
        message:
          type: string
        newBalance:
          type: integer
      required:
        - coins
        - message
        - newBalance
    StreakResponse:
      type: object
      properties:
        streakCount:
          type: integer
        isNewDay:
          type: boolean
        message:
          type: string
      required:
        - streakCount
        - isNewDay
        - message
...
    GenerateLessonBody:
      type: object
      properties:
        subject:
          type: string
          enum: [math, french, spanish, maltese, italian, grammar, history, geography]
        exerciseType:
          type: string
        level:
          type: integer
          minimum: 1
          maximum: 12
        language:
          type: string
          nullable: true
      required:
        - subject
        - exerciseType
        - level
    Lesson:
      type: object
      properties:
        id:
          type: string
        subject:
          type: string
        exerciseType:
          type: string
        level:
          type: integer
        title:
          type: string
        content:
          type: string
        questions:
          type: array
          items:
            $ref: "#/components/schemas/LessonQuestion"
        audioText:
          type: string
          nullable: true
      required:
        - id
        - subject
        - exerciseType
        - level
        - title
        - content
        - questions
        - audioText
    LessonQuestion:
      type: object
      properties:
        id:
          type: string
        question:
          type: string
        type:
          type: string
          enum: [multiple_choice, fill_blank, match, write, speak]
        options:
          type: array
          items:
            type: string
          nullable: true
        correctAnswer:
          type: string
        pairs:
          type: array
          items:
            $ref: "#/components/schemas/MatchPair"
          nullable: true
      required:
        - id
        - question
        - type
        - options
        - correctAnswer
        - pairs
    MatchPair:
      type: object
      properties:
        left:
          type: string
        right:
          type: string
      required:
        - left
        - right
    TTSBody:
      type: object
      properties:
        text:
          type: string
        voice:
          type: string
          enum: [alloy, echo, fable, onyx, nova, shimmer]
          nullable: true
      required:
        - text
    TTSResponse:
      type: object
      properties:
        audio:
          type: string
          description: Base64-encoded MP3 audio
      required:
        - audio
    TranscribeBody:
      type: object
      properties:
        audio:
          type: string
          description: Base64-encoded audio data
      required:
        - audio
    TranscribeResponse:
      type: object
      properties:
        text:
          type: string
      required:
        - text
...