# lib/api-zod/src/generated/types/tTSBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Bloom Learning App API
 * OpenAPI spec version: 0.1.0
 */
import type { TTSBodyVoice } from "./tTSBodyVoice";

export interface TTSBody {
  text: string;
  voice?: TTSBodyVoice;
}